Create Medidect India
Breast cancer is a complex disease that requires prompt diagnosis and effective treatment. As medical science continues to evolve, advancements in breast cancer treatment and detection methods are transforming the landscape of breast cancer care. In this article, we will explore the latest developments in breast cancer treatment, including targeted therapies and personalized medicine. Additionally, we will delve into the role of blood tests in the detection of breast cancer, examining their potential as non-invasive screening tools for early diagnosis.